Making sure the refresher keeps refreshing
Fixes an issue where it was possible for the refreshing timestamp to
get deleted and never get reset when the refreshTimestamps expirer
would call getRemoteValue, and then there were any issues reading from
the cache or from the remote server.

Now, resetting the refresh timestamp expirer is always the first thing
done in the getRemoteValue function.